--- izzet/assets/static.go
@@ -1,41 +1,0 @@
-package assets
-
-import (
-	"fmt"
-
-	"github.com/kkevinchou/izzet/izzet/assets/fonts"
-	"github.com/kkevinchou/izzet/izzet/assets/loaders"
-	"github.com/kkevinchou/izzet/izzet/assets/textures"
-	"github.com/kkevinchou/kitolib/modelspec"
-)
-
-func (a *AssetManager) GetTexture(name string) *textures.Texture {
-	if _, ok := a.textures[name]; !ok {
-		panic(fmt.Sprintf("could not find texture %s", name))
-	}
-	return a.textures[name]
-}
-
-func (a *AssetManager) GetDocument(name string) *modelspec.Document {
-	if _, ok := a.documents[name]; !ok {
-		panic(fmt.Sprintf("could not find animated model %s", name))
-	}
-	return a.documents[name]
-}
-
-func (a *AssetManager) GetFont(name string) fonts.Font {
-	if _, ok := a.fonts[name]; !ok {
-		panic(fmt.Sprintf("could not find font %s", name))
-	}
-	return a.fonts[name]
-}
-
-func (a *AssetManager) LoadDocument(name string, filepath string) bool {
-	scene := loaders.LoadDocument(name, filepath)
-	if _, ok := a.documents[name]; ok {
-		fmt.Printf("warning, document with name %s already previously loaded\n", name)
-	}
-
-	a.documents[name] = scene
-	return true
-}
